Edit Button Added
Created a new file called EditModal.js
Basically, what I'm trying to do is when onClick happens in the edit button, the function showEditModal will return the EditModal component and it will render on screen. Right now the problem is the modal is not rendering and its 5AM and I don't know why. It doesn't impact any existing thing. Everything else is already written in.